Added check balance function recode-init.py (commented)
As with all my code, I am not sure of the names for many of the variables, I will try to correct as many as I can, but It's likely that I won't get all of them.  If there is something additional that needs to be implemented for the functions to work, I am unaware of it.  Help on that front would be appreciated or alternatively a heads up on what else would need to be present.